Use a single command to remove all computational results
